05_hackathon_api/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
